Support for --with-hdf5 option to configure
This was previously documented in INSTALL but not actually supported.

I am fairly sure that building CMOR doesn't actually need the HDF5
libraries, but it may include the headers, so it needs to get the
right ones.  Building programs which link with CMOR may well need the
libraries, and will also need to get the right ones.

Files changed

- configure.ac: add support (see note 1 below)
- (configure is generated from configure.ac and committed, see note 3 below)
- Makefile.in: add support (see note 2 below)
- compile_line.tex.in: add support
- setup.py.in: add support

Note 1. configure has a bunch of assignments setting LIB="" which I
think are spurious and which I've removed.

Note 2. Makefile.in is not in a good state: this change hasn't made
things *worse* but I haven't tried to fix any of it.

Note 3. The version of autotools I gave seems to be older than the
one that CMOR's configure was previously built with: it might be worth
recreating it with a newer one.