Use OS UUID library
Rather than rely on the OSSP uuid library, this relies on the standard
Linux one (part of util-linux / util-linux-ng): this has a more
limited interface, but it's still adequate for CMOR's purposes.  This
should mean that CMOR has one less dependency.

This change has not been very well tested yet.