Furtherinfourl (#219)
* Added extra_link_args to Extension arguments

extra_link_args -Wl,-rpath,/path/to/shared/libs are passed by configure, but not used to build the python extension. As a result, the resulting .so binaries have no rpath to find shared libraries, and environmental LD_LIBRARY_PATH must be defined to locate them. This fix avoid the need to set LD_LIBRARY_PATH.

*  fix further_info_url REGEX template