Cmor 3.2.7 (#232)
* Added extra_link_args to Extension arguments

extra_link_args -Wl,-rpath,/path/to/shared/libs are passed by configure, but not used to build the python extension. As a result, the resulting .so binaries have no rpath to find shared libraries, and environmental LD_LIBRARY_PATH must be defined to locate them. This fix avoid the need to set LD_LIBRARY_PATH.

* update cmor.h to 3.2.7

* add recipes and env yaml file

* fix cmor message and test

* revert cmor_variable changes